+++ src/main/java/org/folio/services/holding/HoldingsService.java
@@ -87,6 +87,7 @@ public class HoldingsService {
   }
 
   public Future<Response> createHolding(HoldingsRecord entity) {
+    entity.setEffectiveLocationId(calculateEffectiveLocation(entity));
     return setHrid(entity)
       .compose(hr -> {
         final Promise<Response> postResponse = promise();
@@ -100,6 +101,8 @@ public class HoldingsService {
   }
 
   private Future<Response> updateHolding(HoldingsRecord oldHoldings, HoldingsRecord newHoldings) {
+    newHoldings.setEffectiveLocationId(calculateEffectiveLocation(newHoldings));
+    log.info("updated location: " + newHoldings.getEffectiveLocationId());
     return refuseIfHridChanged(oldHoldings, newHoldings)
       .compose(notUsed -> {
         final Promise<List<Item>> overallResult = promise();
@@ -131,6 +134,11 @@ public class HoldingsService {
   }
 
   public Future<Response> createHoldings(List<HoldingsRecord> holdings, boolean upsert) {
+    
+    for (HoldingsRecord record : holdings) {
+      record.setEffectiveLocationId(calculateEffectiveLocation(record));
+    }
+
     @SuppressWarnings("all")
     final List<Future> setHridFutures = holdings.stream()
       .map(this::setHrid)
@@ -148,6 +156,17 @@ public class HoldingsService {
         return postSyncResult.future()
           .compose(domainEventPublisher.publishCreatedOrUpdated(batchOperation));
       });
+  }
+
+  private String calculateEffectiveLocation(HoldingsRecord record) {
+    String permanentLocationId = record.getPermanentLocationId();
+    String temporaryLocationId = record.getTemporaryLocationId();
+
+    if (temporaryLocationId != null) {
+      return temporaryLocationId;
+    } else {
+      return permanentLocationId;
+    }
   }
 
   private <T> Handler<AsyncResult<T>> handleTransaction(